+++ Changelog.md
@@ -19,6 +19,8 @@ and this project adheres to
 - New targets (**untested**):
  - Apple II (TARGET\_APPLE2)
  - Apple II enhanced (TARGET\_APPLE2ENH)
+- New targets (**not working**):
+ - Atari XL (TARGET\_ATARIXL) - Need adjusted ld65 config.
 
 ### TUNIX Lisp